requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
pytest>=7.0.0
httpx>=0.24.0
motor>=3.3.0
python-dotenv>=1.0.0
mongomock-motor>=0.0.32
